requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
werkzeug==3.0.1